# apis/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "uuid-string",
                    "email": "john.doe@example.com",
                    "nom": "Doe",
                    "prenom": "John",
                    "role": "PATIENT",
                    "actif": true,
                    "dateCreation": "2024-01-01T00:00:00.000Z",
                    "dateModification": "2024-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user (Admin only)",
        "tags": [
          "api/users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search keyword to filter users by email, nom, or prenom",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter users by role",
            "schema": {
              "type": "string",
              "enum": [
                "KINESITHERAPEUTE",
                "ADMIN"
              ]
            }
          },
          {
            "name": "actif",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "uuid-string",
                        "email": "john.doe@example.com",
                        "nom": "Doe",
                        "prenom": "John",
                        "role": "PATIENT",
                        "actif": true,
                        "dateCreation": "2024-01-01T00:00:00.000Z",
                        "dateModification": "2024-01-01T00:00:00.000Z"
                      }
                    ],
                    "meta": {
                      "total": 1,
                      "page": 1,
                      "limit": 10,
                      "totalPages": 1
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get paginated list of users with optional filters",
        "tags": [
          "api/users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "uuid-string",
                    "email": "john.doe@example.com",
                    "nom": "Doe",
                    "prenom": "John",
                    "role": "PATIENT",
                    "actif": true,
                    "dateCreation": "2024-01-01T00:00:00.000Z",
                    "dateModification": "2024-01-01T00:00:00.000Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific user by ID",
        "tags": [
          "api/users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "api/users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "api/users"
        ]
      }
    },
    "/api/users/{id}/change-password": {
      "patch": {
        "operationId": "UsersController_changePassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Invalid current password"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change user password",
        "tags": [
          "api/users"
        ]
      }
    },
    "/api/users/{id}/toggle-status": {
      "patch": {
        "operationId": "UsersController_toggleStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID (UUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User status toggled successfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle user active status",
        "tags": [
          "api/users"
        ]
      }
    },
...
  },
  "info": {
    "title": "KINE-SAAS API",
    "description": "API documentation for the KINE-SAAS physiotherapy backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "Unique email address"
          },
          "password": {
            "type": "string",
            "example": "SecurePassword123!",
            "description": "Password must be at least 6 characters",
            "minLength": 6
          },
          "nom": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "prenom": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "role": {
            "type": "string",
            "enum": [
              "KINESITHERAPEUTE",
              "ADMIN"
            ],
            "example": "KINESITHERAPEUTE",
            "description": "User role in the system"
          },
          "actif": {
            "type": "boolean",
            "example": true,
            "description": "Whether the user account is active",
            "default": true
          }
        },
        "required": [
          "email",
          "password",
          "nom",
          "prenom",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "actif": {
            "type": "boolean",
            "example": true,
            "description": "Whether the user account is active"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "OldPassword123!",
            "description": "Current password"
          },
          "newPassword": {
            "type": "string",
            "example": "NewPassword123!",
            "description": "New password (minimum 6 characters)",
            "minLength": 6
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  }
}